김형중 edit ; 유저 컴포넌트, 페이지 정리
